// src/models/Client.ts
export interface Client{
  "C_CLIENT": string,
  "D_CLIENT": string;
  "REP_CLIENT": string;
  "SECTEUR": string;
  "ADRESSE": string;
  "ADRESSE2": string;
  "TYPECLIENT": string;
  "ADRESSEL1": string;
  "ADRESSEL2": string;
  "PAYS": string;
  "VILLE": string;
  "CODEPOSTAL": string;
  "TEL1": string;
  "TEL2": string;
  "TEL3": string;
  "FAX": string;
  "EMAIL": string;
  "URL": string;
  "REGISTRE": string;
  "ARTICLIMPO": string;
  "IMMATFISC": string;
  "FORME": string;

}